fix another float->enum conversion error with g++ 4
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5752 d7cf8633-e32d-0410-b094-e92efae38249

+++ prvm_cmds.c
@@ -2682,7 +2682,7 @@ void VM_cin_setstate( void )
 	name = PRVM_G_STRING( OFS_PARM0 );
 	VM_CheckEmptyString( name );
 
-	state = (clvideostate_t)PRVM_G_FLOAT( OFS_PARM1 );
+	state = (clvideostate_t)((int)PRVM_G_FLOAT( OFS_PARM1 ));
 
 	video = CL_GetVideo( name );
 	if( video && state > CLVIDEO_UNUSED && state < CLVIDEO_STATECOUNT )
